backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Telegram Bot
aiogram==3.4.1

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1

# Blockchain / Web3
web3==6.15.1
eth-account==0.11.0

# Exchange APIs
python-binance==1.0.19
ccxt==4.2.27

# Background Tasks
celery[redis]==5.3.6
flower==2.0.1

# Security
cryptography==42.0.2
pyotp==2.9.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
aiohttp==3.9.3
python-dotenv==1.0.1
orjson==3.9.13

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0
